+++ tests/test.js
@@ -3941,6 +3941,24 @@
                     this.clock.tick(10000);
                     same(stored_requests.length, 2, "both events should still be in localStorage");
                 });
+                
+                test('requests are not cleared from localStorage when offline', 3, function() {
+                    var onlineStub = sinon.stub(window.navigator, 'onLine').value(false);
+
+                    mixpanel.batchtest.track('storagetest 1');
+                    mixpanel.batchtest.track('storagetest 2');
+                    var stored_requests = JSON.parse(localStorage.getItem(LOCALSTORAGE_EVENTS_KEY));
+                    same(stored_requests.length, 2, "both events should be in localStorage");
+
+                    this.clock.tick(5000);
+                    this.requests[0].respond(0, {}, '');
+
+                    stored_requests = JSON.parse(localStorage.getItem(LOCALSTORAGE_EVENTS_KEY));
+                    same(stored_requests.length, 2, "both events should still be in localStorage");
+                    this.clock.tick(10000);
+                    same(stored_requests.length, 2, "both events should still be in localStorage");
+                    onlineStub.restore();
+                });
 
                 test('orphaned data in localStorage gets sent on init', 6, function() {
                     clearLibInstance(mixpanel.batchtest);
@@ -5243,7 +5261,7 @@
                     }
                 }
 
-                function makeFakeFetchResponse(status, body, cb) {
+                function makeFakeFetchResponse(status, body) {
                     body = body || {}
                     var response = new Response({}, {
                         status: status,
@@ -5262,11 +5280,19 @@
                     );
                 }
 
+                function makeDelayedFetchResponse(status, body, delay) {
+                    return new Promise(function(resolve) {
+                        setTimeout(function() {
+                            resolve(makeFakeFetchResponse(status, body));
+                        }, delay);
+                    });
+                }
+
                 module('recorder', {
                     setup: function () {
                         this.token = rand_name();
-
-                        this.clock = sinon.useFakeTimers();
+                        this.startTime = 1723733423402;
+                        this.clock = sinon.useFakeTimers(this.startTime);
                         this.randomStub = sinon.stub(Math, 'random');
                         this.fetchStub = sinon.stub(window, 'fetch');
 
@@ -5543,7 +5569,6 @@
                         }, this));
                     });
                 });
-
                 asyncTest('retries record request after a 500', 17, function () {
                     this.randomStub.returns(0.02);
                     this.initMixpanelRecorder({record_sessions_percent: 10});
@@ -5577,6 +5602,53 @@
                         same(urlParams.get("seq"), "1", "2nd sequence is retried");
 
                         mixpanel.recordertest.stop_session_recording();
+                    });
+                });
+
+                asyncTest('retries record requests when offline', 7, function () {
+                    var onlineStub = sinon.stub(window.navigator, 'onLine').value(false);
+                    // pretend we can't compress so we can compare the events in the fetch request
+                    var compressionStreamStub = sinon.stub(window, 'CompressionStream').value(undefined);
+                    
+                    this.randomStub.returns(0.02);
+                    this.initMixpanelRecorder({record_sessions_percent: 10});
+                    this.assertRecorderScript(true);
+                    
+                    this.responseBlobStub = sinon.stub(window.Response.prototype, 'blob');
+                    this.responseBlobStub.returns(fakePromiseWrap(new Blob()));
+                    this.fetchStub.onCall(0)
+                        .returns(new Promise(function (_resolve, reject) {
+                            // simulate offline
+                            reject(new TypeError('Failed to fetch'));
+                        })) 
+                        .onCall(1)
+                        .returns(makeFakeFetchResponse(200))
+
+                    this.afterRecorderLoaded.call(this, function () {
+                        simulateMouseClick(document.body);
+                        this.clock.tick(10 * 1000)
+                        same(this.fetchStub.getCalls().length, 1, 'one batch fetch request made every ten seconds');
+
+                        var urlParams = validateAndGetUrlParams(this.fetchStub.getCall(0));
+                        same(urlParams.get("seq"), "0", "first sequence fails because we're offline");
+                        var fetchBody = this.fetchStub.getCall(0).args[1].body;
+
+                        this.clock.tick(20 * 1000);
+
+                        stop();
+                        untilDone(_.bind(function (done) {
+                            var fetchCall1 = this.fetchStub.getCall(1);
+                            if (fetchCall1) {
+                                urlParams = validateAndGetUrlParams(fetchCall1);
+                                same(urlParams.get("seq"), "0", "first sequence is retried with exponential backoff");
+                                same(fetchBody, fetchCall1.args[1].body, 'fetch body should be the same as the first request');
+                            }
+                            onlineStub.restore();
+                            compressionStreamStub.restore();
+                            mixpanel.recordertest.stop_session_recording();
+
+                            done();
+                        }, this))
                     });
                 });
 
@@ -5640,6 +5712,31 @@
                     });
                 });
 
+                asyncTest('respects minimum session length setting', function () {
+                    this.randomStub.returns(0.02);
+                    this.initMixpanelRecorder({record_sessions_percent: 10, record_min_ms: 8000});
+                    this.assertRecorderScript(true)
+
+                    this.afterRecorderLoaded.call(this, function () {
+                        simulateMouseClick(document.body);
+                        this.clock.tick(5 * 1000);
+
+                        mixpanel.recordertest.stop_session_recording();
+
+                        stop();
+                        untilDone(_.bind(function(done) {
+                            same(this.fetchStub.getCalls().length, 0, 'does not flush events if session is too short');
+
+                            simulateMouseClick(document.body);
+                            this.clock.tick(20 * 1000);
+                            realSetTimeout(_.bind(function() {
+                                same(this.fetchStub.getCalls().length, 0, 'no fetch requests after recording is stopped');
+                                done();
+                            }, this), 2);
+                        }, this));
+                    });
+                });
+
                 asyncTest('resets after idle timeout', 14, function () {
                     this.randomStub.returns(0.02);
                     this.initMixpanelRecorder({record_sessions_percent: 10});
@@ -5661,7 +5758,11 @@
                         same(urlParams.get('seq'), '0', 'sends first sequence');
                         var replayId1 = urlParams.get('replay_id');
 
-                        this.clock.tick(33 * 60 * 1000);
+                        this.clock.tick(16 * 60 * 1000);
+                        // simulate a mutation event to ensure we don't try sending it until there's a user event
+                        document.body.appendChild(document.createElement('div'));
+                        this.clock.tick(16 * 60 * 1000);
+
                         same(this.fetchStub.getCalls().length, 1, 'no new record requests after idle timeout');
 
                         simulateMouseClick(document.body);
@@ -5673,6 +5774,64 @@
                         ok(replayId1 !== replayId2, 'replay id is different after reset');
 
                         mixpanel.recordertest.stop_session_recording();
+                    });
+                });
+
+
+                asyncTest('handles race condition where the recording resets while a request is in flight', 14, function () {
+                    this.randomStub.returns(0.02);
+
+                    this.initMixpanelRecorder({record_sessions_percent: 10});
+                    this.assertRecorderScript(true);
+                    this.randomStub.restore(); // restore the random stub after script is loaded for batcher uuid dedupe
+
+                    this.responseBlobStub = sinon.stub(window.Response.prototype, 'blob');
+                    this.responseBlobStub.returns(fakePromiseWrap(new Blob()));
+                    this.fetchStub.onFirstCall()
+                        .returns(makeDelayedFetchResponse(200, {}, 5 * 1000))
+
+                    this.fetchStub.callThrough();
+
+                    this.afterRecorderLoaded.call(this, function () {
+                        // current replay stops, a new one starts a bit after
+                        mixpanel.recordertest.stop_session_recording();
+
+                        same(this.fetchStub.getCalls().length, 1, 'flushed events after stopping recording');
+                        var urlParams = validateAndGetUrlParams(this.fetchStub.getCall(0));
+                        same(urlParams.get('seq'), '0', 'sends first sequence');
+                        
+                        if (!IS_RECORDER_BUNDLED) {
+                            same(urlParams.get('replay_start_time'), (this.startTime / 1000).toString(), 'sends the right start time');
+                        } else {
+                            ok(true, 'cannot test replay_start_time when recorder is bundled, rrweb stores a reference to Date.now at the global level so stubs / fake timers will not work.');
+                        }
+
+                        var replayId1 = urlParams.get('replay_id');
+
+                        // start recording again while the first replay's request is in flight
+                        mixpanel.recordertest.start_session_recording();
+                        simulateMouseClick(document.body);
+                        stop();
+
+                        untilDone(_.bind(function (done) {
+                            this.clock.tick(10 * 1000);
+                            var fetchCall1 = this.fetchStub.getCall(1);
+                            if (fetchCall1) {
+                                urlParams = validateAndGetUrlParams(this.fetchStub.getCall(1));
+                                same(urlParams.get('seq'), '0', 'new replay uses the first sequence');
+
+                                if (!IS_RECORDER_BUNDLED) {
+                                    same(urlParams.get('replay_start_time'), (this.startTime / 1000).toString(), 'sends the right start time');
+                                } else {
+                                    ok(true, 'cannot test replay_start_time when recorder is bundled, rrweb stores a reference to Date.now at the global level so stubs / fake timers will not work.');
+                                }
+
+                                var replayId2 = urlParams.get('replay_id');
+                                ok(replayId1 !== replayId2, 'replay id is different after reset');
+                                mixpanel.recordertest.stop_session_recording();
+                                done();
+                            }
+                        }, this));
                     });
                 });
             }